FORK: Removes old Node/Resp representation in dgraph
This bring to specialize the node struct. Now the tx node is exported and from here should start the refactoring that adapts the logic of the whole program to use the tx node. FORK: save data to generate tx from the node and export that of adapt all all heuristics to use tx node?